+++ src/pages/Store/StoreDetailPage/index.tsx
@@ -5,21 +5,11 @@ import cn from 'utils/ts/classnames';
 import * as api from 'api';
 import styles from './StoreDetailPage.module.scss';
 
-const useStoreDetail = (params: any) => {
-  const { data: storeDetail } = useQuery(
-    ['storeDetail', params],
-    ({ queryKey }) => api.store.getStoreDetailInfo(queryKey[1]),
-    { retry: 0 },
-  );
-
-  return storeDetail;
-};
-
 function StoreDetailPage() {
   const params = useParams();
   const isMobile = useMediaQuery();
-  const storeDetailInfo = useStoreDetail(params.id);
-  console.log(storeDetailInfo);
+  const { storeDetail, storeDescription } = useStoreDetail(params.id);
+
   return (
     <div className={styles.template}>
       <div className={styles.section}>
